Refactor: multiple aspects of morph/fill code
Change input order for tile 9-grids from
center, orthogonal, diagonal (clockwise, starting n/ne) to
left-to-right, top-to-bottom.

Remove templating from ninegrid initialization.

Split up floodfill/morph C++ code more. New fill_common.*pp
files for shared data structures and components instead
of having morphology header include floodfill header.

New threadsafe static PixelBuffer creation.

Rename constant tile class to the more intuitive ConstTiles,
shortening X_ALPHA_TILE to ALPHA_X for X=transparent/opaque

Some macro fixes
