+++ lib/mypaintlib.hpp
@@ -23,6 +23,7 @@
 #include "colorchanger_crossed_bowl.hpp"
 #include "gdkpixbuf2numpy.hpp"
 #include "fastpng.hpp"
+#include "fill_common.hpp"
 #include "floodfill.hpp"
 #include "morphology.hpp"
 #include "brushsettings.hpp"
