+++ lib/mypaintlib.i
@@ -30,6 +30,8 @@ typedef struct { int x, y, w, h; } Rect;
 %include "colorchanger_wash.hpp"
 %include "colorchanger_crossed_bowl.hpp"
 %include "fastpng.hpp"
+
+%include "fill_common.hpp"
 %include "floodfill.hpp"
 %include "morphology.hpp"
 %include "brushsettings.hpp"
